# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chat API",
    "description": "A little whatsapp api clone",
    "contact": "jefsteven22@gmail.com",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/users/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login User",
        "operationId": "LoginUser",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJpZCI6OSwiZmlyc3RuYW1lIjoiU3RldmVuIiwibGFzdG5hbWUiOiJDb250cmVyYXMiLCJlbWFpbCI6InN0ZXZlbjJAbWFpbC5jb20iLCJhdmF0YXIiOiJodHRwczovL2Nkbi5waXhhYmF5LmNvbS9waG90by8yMDE1LzEwLzA1LzIyLzM3L2JsYW5rLXByb2ZpbGUtcGljdHVyZS05NzM0NjBfMTI4MC5wbmciLCJkZXNjcmlwdGlvbiI6bnVsbCwidmFsaWRFbWFpbCI6ZmFsc2UsImNyZWF0ZWRBdCI6IjIwMjMtMDktMTVUMDI6NDc6MDYuMDQzWiIsInVwZGF0ZWRBdCI6IjIwMjMtMDktMTVUMDI6NDc6MDYuMDQzWiIsImlhdCI6MTY5NDgzMTQ2NCwiZXhwIjoxNjk0ODM1MDY0fQ.SoNMl-9zB0Rvdl9k7ie2FKYeTCuAOhfalKoWzHN4-FBHwVW-F-KvQIaXmO3iqPFsfBQ5i3QZv4NpGPSFW0edAg"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LoginUserRequest"
                  },
                  {
                    "example": {
                      "email": "dennise@gmail.com",
                      "password": "Jeffrey24214736."
                    }
                  }
                ]
              },
              "example": {
                "email": "dennise@gmail.com",
                "password": "Jeffrey24214736."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginUserRequest": {
        "title": "LoginUserRequest",
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "dennise@gmail.com",
          "password": "Jeffrey24214736."
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}